CRNT

AbstractAlgebra.Generic.matrix

CRNT.Diagonal

CRNT.Jacobian

CRNT.NonnegativeNullspaceCone

AbstractAlgebra.Generic.matrix — Method

matrix(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl one-column matrix over the given ring R, assuming all the entries can be coerced into R.

CRNT.Diagonal — Method

Diagonal(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl diagonal matrix over the given ring R, assuming all the entries can be coerced into R.

CRNT.Jacobian — Function

Jacobian(R, M)
Jacobian(R, M, xs=1:length(Nemo.gens(R)))

Returns the Jacobian matrix of a one-column matrix of polynomials M with respect to the generators of R indexed by xs. When xs is omitted all the generators of R are used.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> M = matrix(R, [vars[1]*vars[2]*vars[6]-vars[8]; vars[3]*vars[9]+2*vars[7]])
[k1*k2*x1-x3]
[ k3*x4+2*x2]

julia> Jacobian(R, M, 6:9)
[k1*k2  0  -1   0]
[    0  2   0  k3]

julia> Jacobian(R,M)
[k2*x1  k1*x1   0  0  0  k1*k2  0  -1   0]
[    0      0  x4  0  0      0  2   0  k3]

CRNT.NonnegativeNullspaceCone — Method

NonnegativeNullspaceCone(N::AbstractMatrix{T}) where {T<:Integer}

Return a matrix whose columns generate the cone intersection of the nonnegative orthant and the nullsapce of N.

Examples

julia> NonnegativeNullspaceCone([0 0; 0 0])
2×2 Array{Int64,2}:
 1  0
 0  1

julia> NonnegativeNullspaceCone([1 -1; 2 -2])
2×1 Array{Int64,2}:
 1
 1