CRNT

LinearAlgebra.Diagonal

AbstractAlgebra.Generic.matrix

CRNT.Collectallrows

CRNT.Jacobian

CRNT.collect_homoexponent_vectors

CRNT.cone_positivenullspace

CRNT.dissect

CRNT.homoexponent_vectors

CRNT.homogenize

LinearAlgebra.Diagonal — Method

Diagonal(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl diagonal matrix over the given ring R, assuming all the entries can be coerced into R.

AbstractAlgebra.Generic.matrix — Method

matrix(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl one-column matrix over the given ring R, assuming all the entries can be coerced into R.

CRNT.Collectallrows — Method

Collectallrows(predicate, p::MPolyElem, V::AbstractMatrix)

Returns an Array of all the exponents of terms of p for which the exponent is a row of V and the coefficient satisfies predicate.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> q = vars[1]*vars[2]*vars[6]-vars[3]*vars[8]+vars[5]*vars[9]-2*vars[1]*vars[3]*vars[5]^2
k1*k2*x1-2*k1*k3*k5^2-k3*x3+k5*x4

julia> collect(exponent_vectors(q))
4-element Array{Array{Int64,1},1}:
 [1, 1, 0, 0, 0, 1, 0, 0, 0]
 [1, 0, 1, 0, 2, 0, 0, 0, 0]
 [0, 0, 1, 0, 0, 0, 0, 1, 0]
 [0, 0, 0, 0, 1, 0, 0, 0, 1]

julia> V =  [[1 1 0 0 0 1 0 0 0]; [0 0 1 0 0 0 0 1 0]; [1 0 1 0 2 0 0 0 0]];

julia> Collectallrows(x->Nemo.isless(x,0), q, V)
2-element Array{Array{Int64,1},1}:
 [1, 0, 1, 0, 2, 0, 0, 0, 0]
 [0, 0, 1, 0, 0, 0, 0, 1, 0]

CRNT.Jacobian — Function

Jacobian(R, M)
Jacobian(R, M, vars=1:length(gens(R)))

Returns the Jacobian matrix of a one-column matrix of polynomials M with respect to the generators of R indexed by vars. When vars is omitted all the generators of R are used.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> M = [vars[1]*vars[2]*vars[6]-vars[8]; vars[3]*vars[9]+2*vars[7]]
[k1*k2*x1-x3]
[ k3*x4+2*x2]

julia> Jacobian(R, M, 6:9)
[k1*k2  0  -1   0]
[    0  2   0  k3]

julia> Jacobian(R,M)
[k2*x1  k1*x1   0  0  0  k1*k2  0  -1   0]
[    0      0  x4  0  0      0  2   0  k3]

CRNT.collect_homoexponent_vectors — Method

CollectExponent_homovectors(p::MPolyElem)

Collects the homogenized (with an extra first component set to 1) exponents of p in the rows of a matrix. The returned matrix can be used to compute the Newton polytope of p.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> q = vars[1]*vars[2]*vars[6]-vars[8]
k1*k2*x1-x3

julia> collect(exponent_vectors(q))
2-element Array{Array{Int64,1},1}:
 [1, 1, 0, 0, 0, 1, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 1, 0]

julia> CollectExponent_homovectors(q)
2×10 Transpose{Int64,Array{Int64,2}}:
 1  1  1  0  0  0  1  0  0  0
 1  0  0  0  0  0  0  0  1  0

Recall it is approximately ten times faster than the equivalent transpose(reduce(hcat,collect(exponent_homovectors(q))))

CRNT.cone_positivenullspace — Method

cone_positivenullspace(N::AbstractMatrix{T}) where {T<:Integer}

Return a matrix whose columns generate the cone intersection of the nonnegative orthant and the nullsapce of N.

Examples

julia> cone_positivenullspace([0 0; 0 0])
2×2 Array{Int64,2}:
 1  0
 0  1

julia> cone_positivenullspace([1 -1; 2 -2])
2×1 Array{Int64,2}:
 1
 1

CRNT.dissect — Method

dissect(p::MPolyElem)

Iterator for the coefficients and exponent vectors of the given polynomial.

CRNT.homoexponent_vectors — Method

homoexponent_vectors(p::MPolyElem)

An iterator for the homogenized (with an extra first component set to '1') exponent vectors of the multivariate polynomial p. To retrieve an array, use collect(homoexponent_vectors(p)).

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> p = vars[1]*vars[2]*vars[6]-vars[8]
k1*k2*x1-x3

julia> collect(exponent_vectors(p))
2-element Array{Array{Int64,1},1}:
 [1, 1, 0, 0, 0, 1, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 1, 0]

julia> collect(homoexponent_vectors(p))
2-element Array{Any,1}:
 [1, 1, 1, 0, 0, 0, 1, 0, 0, 0]
 [1, 0, 0, 0, 0, 0, 0, 0, 1, 0]

CRNT.homogenize — Method

homogenize(iter)

Given an iterator iter whose values are 1d arrays, returns an iterator that yields [1; x] where 1 is a unit of the element type of x, and x is a value of iter.

Examples

julia> x = [[1, 2], [3, 4], [5, 6]];

julia> collect(homogenize(x))
3-element Array{Array{Int64,1},1}:
 [1, 1, 2]
 [1, 3, 4]
 [1, 5, 6]