CRNT

AbstractAlgebra.Generic.matrix

CRNT.CollectHomoiterInmatrows

CRNT.Diagonal

CRNT.Jacobian

CRNT.NonnegativeNullspaceCone

CRNT.exponent_homovectors

CRNT.homogenize

AbstractAlgebra.Generic.matrix — Method

matrix(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl one-column matrix over the given ring R, assuming all the entries can be coerced into R.

CRNT.CollectHomoiterInmatrows — Method

CollectHomoiterInmatrows(iter)

Given an iterator iter whose values are 1d arrays, collect its homogenized (with an extra first component set to 1) values in the rows of a matrix.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> p = vars[1]*vars[2]*vars[6]-vars[8]
k1*k2*x1-x3

julia> collect(exponent_vectors(p))
2-element Array{Array{Int64,1},1}:
 [1, 1, 0, 0, 0, 1, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 1, 0]

julia> CollectHomoiterInmatrows(exponent_vectors(p))
2×10 Transpose{Int64,Array{Int64,2}}:
 1  1  1  0  0  0  1  0  0  0
 1  0  0  0  0  0  0  0  1  0

Recall it is approximately ten times more efficient than the equivalent transpose(reduce(hcat,collect(exponent_homovectors(q))))

CRNT.Diagonal — Method

Diagonal(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl diagonal matrix over the given ring R, assuming all the entries can be coerced into R.

CRNT.Jacobian — Method

Jacobian(R, M)
Jacobian(R, M, vars=1:length(Nemo.gens(R)))

Returns the Jacobian matrix of a one-column matrix of polynomials M with respect to the generators of R indexed by vars. When vars is omitted all the generators of R are used.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> M = matrix(R, [vars[1]*vars[2]*vars[6]-vars[8]; vars[3]*vars[9]+2*vars[7]])
[k1*k2*x1-x3]
[ k3*x4+2*x2]

julia> Jacobian(R, M, 6:9)
[k1*k2  0  -1   0]
[    0  2   0  k3]

julia> Jacobian(R,M)
[k2*x1  k1*x1   0  0  0  k1*k2  0  -1   0]
[    0      0  x4  0  0      0  2   0  k3]

CRNT.NonnegativeNullspaceCone — Method

NonnegativeNullspaceCone(N::AbstractMatrix{T}) where {T<:Integer}

Return a matrix whose columns generate the cone intersection of the nonnegative orthant and the nullsapce of N.

Examples

julia> NonnegativeNullspaceCone([0 0; 0 0])
2×2 Array{Int64,2}:
 1  0
 0  1

julia> NonnegativeNullspaceCone([1 -1; 2 -2])
2×1 Array{Int64,2}:
 1
 1

CRNT.exponent_homovectors — Method

exponent_homovectors(p::MPolyElem)

An iterator for the homogenized (with an extra first component set to 1) exponent vectors of the multivariate polynomial p. To retrieve an array, use collect(exponent_homovectors(p)).

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> p = vars[1]*vars[2]*vars[6]-vars[8]
k1*k2*x1-x3

julia> collect(exponent_vectors(p))
2-element Array{Array{Int64,1},1}:
 [1, 1, 0, 0, 0, 1, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 1, 0]

julia> collect(exponent_homovectors(p))
2-element Array{Any,1}:
 [1, 1, 1, 0, 0, 0, 1, 0, 0, 0]
 [1, 0, 0, 0, 0, 0, 0, 0, 1, 0]

CRNT.homogenize — Method

homogenize(iter)

Given an iterator iter whose values are 1d arrays, returns an iterator that yields [1; x] where 1 is a unit of the element type of x, and x is a value of iter.

Examples

julia> x = [[1, 2], [3, 4], [5, 6]];

julia> collect(homogenize(x))
3-element Array{Array{Int64,1},1}:
 [1, 1, 2]
 [1, 3, 4]
 [1, 5, 6]